Adds functions for tabulating feature properties
