+++ phot_class/__init__.py
@@ -9,3 +9,7 @@ results for an entire photometric survey.
 """
 
 # Todo: Include an example(s)
+
+from pathlib import Path as _Path
+
+dust_path = _Path(__file__).resolve().parent / 'schlegel98_dust_map'
